Trix/fedora upstream (#1426)
* Check if this rpmdevtools is installed

rpmdev-setuptree is provided by rpmdevtools.
So check that rpmdevtools is installed before trying to use.

* Check that rpmbuild succeeds before continuing.

rpmbuild will fail if the prerequisites of the opae.spec are not met.
So report the error and stop.

+++ scripts/create-rpms.sh
@@ -7,6 +7,13 @@ if [ $? -eq 1 ]; then
 	exit 1
 fi
 
+#check if rpmdevtools is installed
+yum list installed rpmdevtools
+if [ $? -eq 1 ]; then
+	echo "'rpmdevtools' package not installed.. exiting"
+	exit 1
+fi
+
 rm -rf ~/rpmbuild
 rpmdev-setuptree
 
@@ -57,7 +64,16 @@ cd ~/rpmbuild/SPECS/
 
 #generate RPMS
 rpmbuild -ba opae.spec
+if [ $? -eq 1 ]; then
+	echo "building rpm failed.. exiting"
+	exit 1
+fi
+
 rpmbuild -bs opae.spec
+if [ $? -eq 1 ]; then
+	echo "building srpm failed.. exiting"
+	exit 1
+fi
 
 #copy RPMS to build directory
 cp ~/rpmbuild/RPMS/x86_64/opae-* $BUILD_DIR/
